requirements.txt
uvicorn>=0.30.1
fastapi>=0.111.0
SQLAlchemy>=2.0.31
pydantic>=2.7.4
pydantic_core>=2.18.4
datedelta>=1.4
tqdm>=4.66.4
requests>=2.31.0
nltk>=3.8.1
joblib>=1.4.2
numpy>=2.0.0
scikit-learn>=1.5.0
keras>=3.4.1
bs4>=0.0.2
beautifulsoup4>=4.12.3
starlette>=0.37.2
selenium>=4.22.0
PyYAML>=6.0.1
wheel>=0.42.0
setuptools>=69.1.1
typing_extensions>=4.12.2
packaging>=24.1
